:sparkles: Implemented a high-level USB interface class made to represent USB features outside of base configuration. For example, an interface could be a speaker, keyboard, mouse, or serial device. This is intended to be used at the level below the configuration descriptor. Which can either be an interface descriptor or an interface assoication descriptor. This interface is expected to manage its own endpoints and interact with the control endpoint (EP0) via callbacks. usb.hpp which previously had the endpoint declarations now is just a shortcut for including interface.hpp and endpoint.hpp. The endpoints are now in endpoint.hpp
