+++ tests/usb.test.cpp
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 #include "libhal/scatter_span.hpp"
+#include "mock_usb_endpoints.hpp"
 #include <libhal/usb.hpp>
 
 #include <libhal/error.hpp>
@@ -21,173 +22,7 @@
 #include <boost/ut.hpp>
 
 namespace hal::v5 {
-namespace {
-// Mock implementation for usb_endpoint
 
-class mock_usb_endpoint : public usb_endpoint
-{
-public:
-  usb_endpoint_info m_info{};
-  bool m_stall_called{ false };
-  bool m_should_stall{ false };
-  bool m_reset_called{ false };
-
-protected:
-  [[nodiscard]] usb_endpoint_info driver_info() const override
-  {
-    return m_info;
-  }
-
-  void driver_stall(bool p_should_stall) override
-  {
-    m_stall_called = true;
-    m_should_stall = p_should_stall;
-  }
-
-  void driver_reset() override
-  {
-    m_reset_called = true;
-  }
-};
-
-// Mock implementation for usb_control_endpoint
-class mock_usb_control_endpoint : public usb_control_endpoint
-{
-public:
-  mock_usb_endpoint m_endpoint;
-  bool m_should_connect{ false };
-  u8 m_address{ 0 };
-  scatter_span<byte const> m_write_data{};
-  scatter_span<byte> m_read_buffer{};
-  usize m_read_result{ 0 };
-  callback<void(on_receive_tag)> m_receive_callback{};
-  bool m_on_receive_called{ false };
-
-private:
-  [[nodiscard]] usb_endpoint_info driver_info() const override
-  {
-    return m_endpoint.info();
-  }
-
-  void driver_stall(bool p_should_stall) override
-  {
-    m_endpoint.stall(p_should_stall);
-  }
-
-  void driver_connect(bool p_should_connect) override
-  {
-    m_should_connect = p_should_connect;
-  }
-
-  void driver_set_address(u8 p_address) override
-  {
-    m_address = p_address;
-  }
-
-  void driver_write(scatter_span<byte const> p_data) override
-  {
-    m_write_data = p_data;
-  }
-
-  usize driver_read(scatter_span<byte> p_buffer) override
-  {
-    m_read_buffer = p_buffer;
-    return m_read_result;
-  }
-
-  void driver_on_receive(
-    callback<void(on_receive_tag)> const& p_callback) override
-  {
-    m_receive_callback = p_callback;
-    m_on_receive_called = true;
-  }
-
-  void driver_reset() override
-  {
-    m_endpoint.reset();
-  }
-};
-
-// Mock implementation for usb_in_endpoint
-class mock_usb_in_endpoint : public usb_in_endpoint
-{
-public:
-  mock_usb_endpoint m_endpoint;
-  scatter_span<byte const> m_write_data{};
-
-private:
-  [[nodiscard]] usb_endpoint_info driver_info() const override
-  {
-    return m_endpoint.info();
-  }
-
-  void driver_stall(bool p_should_stall) override
-  {
-    m_endpoint.stall(p_should_stall);
-  }
-  void driver_write(scatter_span<byte const> p_data) override
-  {
-    m_write_data = p_data;
-  }
-
-  void driver_reset() override
-  {
-    m_endpoint.reset();
-  }
-};
-
-// Mock implementation for usb_out_endpoint
-class mock_usb_out_endpoint : public usb_out_endpoint
-{
-public:
-  mock_usb_endpoint m_endpoint;
-  callback<void(on_receive_tag)> m_receive_callback{};
-  bool m_on_receive_called{ false };
-  scatter_span<byte> m_read_buffer{};
-  usize m_read_result{ 0 };
-
-private:
-  [[nodiscard]] usb_endpoint_info driver_info() const override
-  {
-    return m_endpoint.info();
-  }
-
-  void driver_stall(bool p_should_stall) override
-  {
-    m_endpoint.stall(p_should_stall);
-  }
-
-  void driver_on_receive(
-    callback<void(on_receive_tag)> const& p_callback) override
-  {
-    m_receive_callback = p_callback;
-    m_on_receive_called = true;
-  }
-
-  usize driver_read(scatter_span<byte> p_buffer) override
-  {
-    m_read_buffer = p_buffer;
-    return m_read_result;
-  }
-
-  void driver_reset() override
-  {
-    m_endpoint.reset();
-  }
-};
-
-struct mock_usb_interrupt_in_endpoint : public mock_usb_in_endpoint
-{};
-
-struct mock_usb_bulk_in_endpoint : public mock_usb_in_endpoint
-{};
-
-struct mock_usb_interrupt_out_endpoint : public mock_usb_out_endpoint
-{};
-
-struct mock_usb_bulk_out_endpoint : public mock_usb_out_endpoint
-{};
-}  // namespace
 // Test for usb_endpoint_info methods
 boost::ut::suite<"usb_endpoint_info_test"> endpoint_info_test = []() {
   using namespace boost::ut;
